refactor(cli): extract diagnostic commands
